complete update user
