+++ src/main/java/com/project/controller/user/BaseController.java
@@ -19,15 +19,6 @@ import javax.validation.Valid;
 @Controller
 public class BaseController {
 
-    @Autowired
-    private UserService userService;
-
-    @Autowired
-    private RoleService roleService;
-
-    @Autowired
-    private SendMail sendMail;
-
     @GetMapping("/login")
     public String getPageLogin(){
         return "user/login";
@@ -37,51 +28,6 @@ public class BaseController {
     public String getPageRegister(Model model){
         model.addAttribute("user", new User());
         return "user/register";
-    }
-
-    @PostMapping("/create-user")
-    public String doRegister(@ModelAttribute @Valid User user, BindingResult result){
-        if (result.hasErrors()){
-            return "redirect:/register?invalid";
-        }
-
-        User u = userService.save(user);
-        if (u == null)
-            return "redirect:/register?existed";
-
-        Thread thread = new Thread(new Runnable() {
-            @Override
-            public void run() {
-                try{
-                    String url = "<a href=\"http://localhost:8080/actived/" + u.getToken() +
-                            "\">Actice account !</a>";
-                    sendMail.confirmMail(u.getEmail(),url);
-                    Thread.sleep(10*60*1000);
-                    if (userService.existUserByTokenAndIsEnable(u.getToken(), false)){
-                        userService.delete(u.getId());
-                    }
-                }catch (InterruptedException ex){
-                    ex.printStackTrace();
-                }
-            }
-        });
-        thread.start();
-
-        return "redirect:/login";
-    }
-
-    @GetMapping("/actived/{token}")
-    public String activeUser(@PathVariable(value = "token") String token){
-        User user = userService.findUserByTokenAndIsEnable(token, false).orElse(null);
-        if (user != null){
-            //add role for user
-            Role role = roleService.findRoleByRole("USER");
-            user.addRole(role);
-            user.setIsEnable(true);
-            userService.save(user);
-            return "redirect:/login?actived";
-        }
-        return "redirect:/login?expiredToken";
     }
 
     @GetMapping("/contact")
