+++ src/main/java/com/project/repository/UserRepository.java
@@ -11,9 +11,11 @@ public interface UserRepository extends JpaRepository<User, Integer> {
 
     Optional<User> findUserByUserNameAndIsEnable(String userName, Boolean isEnable);
 
-    Boolean existsUserByUserName(String username);
-
     Boolean existsUserByTokenAndIsEnable(String token, Boolean isEnable);
 
     Optional<User> findUserByTokenAndIsEnable(String token, Boolean isEnable);
+
+    Boolean existsUserByUserName(String username);
+
+    Optional<User> findUserByUserName(String username);
 }
