+++ src/main/java/com/project/service/UserService.java
@@ -7,13 +7,17 @@ import java.util.Optional;
 public interface UserService {
     User findUserByUserNameAndIsEnable(String userName, Boolean isEnable);
 
+    Boolean existUserByTokenAndIsEnable(String token, Boolean isEnable);
+
     Optional<User> findUserByTokenAndIsEnable(String token, Boolean isEnable);
 
     User save(User user);
 
+    User activeUser(User user);
+
     Boolean existUserByUserName(String username);
 
-    Boolean existUserByTokenAndIsEnable(String token, Boolean isEnable);
+    Optional<User> findUserByUserName(String username);
 
     void delete(Integer id);
 }
