+++ src/main/resources/templates/user/contact.html
@@ -42,6 +42,11 @@
                         </div>
                         <div class="contact__form">
                             <h5>GỬI LIÊN HỆ</h5>
+
+                            <th:block sec:authorize="isAuthenticated()">
+                                <a th:href="@{/user/{username}(username = ${#authentication.getPrincipal().getUsername()})}">Click vao day</a>
+                            </th:block>
+
                             <form action="#" th:action="@{contact}" th:object="${contact}" method="post">
                                 <input type="hidden" th:field="*{id}">
                                 <input type="text" th:field="*{name}" placeholder="Name">
